requirements.txt
openai
tritonclient[http]
numpy
transformers
pyyaml
jupyter